stream-pipeline/cloud_functions/scrap_twits_jobs/requirements.txt
google-cloud-storage==1.37.0
google-cloud-pubsub==2.4.1
python-dateutil==2.8.1
bs4==0.0.1
twint @ git+https://github.com/twintproject/twint.git@4b30fa33991ccdf89c77471382daad3d3bc8c988
